backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
httpx==0.27.0
pyserial-asyncio==0.6
python-dotenv==1.0.1
bcrypt==4.1.3
anyio==4.3.0
starlette==0.37.2